+++ agrotera-test/src/main/java/net/onedaybeard/agrotera/component/ComponentA.java
@@ -1,0 +1,7 @@
+package net.onedaybeard.agrotera.component;
+
+import com.artemis.Component;
+
+public class ComponentA extends Component {
+
+}