+++ agrotera-test/src/main/java/net/onedaybeard/agrotera/system/BasicSystem.java
@@ -1,0 +1,28 @@
+package net.onedaybeard.agrotera.system;
+
+import static org.junit.Assert.assertNotNull;
+import lombok.ArtemisSystem;
+import net.onedaybeard.agrotera.component.ComponentA;
+import net.onedaybeard.agrotera.component.ComponentB;
+
+import com.artemis.Entity;
+import com.artemis.systems.EntityProcessingSystem;
+
+@ArtemisSystem(
+	requires=ComponentA.class,
+	excludes=ComponentB.class)
+public class BasicSystem extends EntityProcessingSystem {
+
+	public BasicSystem() {
+		super(null);
+	}
+
+	@Override
+	protected void process(Entity e) {
+
+	}
+	
+	public void assertAMapper() {
+		assertNotNull(componentAMapper);
+	}
+}