+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  - IOOB exception during agrotera:matrix if no systems/managers were processed.
  - Java 1.6 support.
+ - **Fix**: requiresOne mappers weren't injected.
 
 ## Version 0.2.1 - 2013-08-08
  - Multithreaded class weaving.